Print the cycle counter without PRIu64
Bill Howland hit this building on Linux Mint and worked out both the
failure and the fix:

  src/rpi/performance.c:226:26: error: expected ')' before 'PRIu64'
  note: 'PRIu64' is defined in header '<inttypes.h>'; did you forget to
        '#include <inttypes.h>'?

That note is a red herring - it is keyed on the identifier name and does
not check whether the header is already there, and it is, on line 4, as
it has been since the counters were added (and as it is upstream).

The real cause is inside newlib. PRIu64 sits in a "#if __int64_t_defined"
block in inttypes.h; that macro comes from sys/_stdint.h, which keys off
the older three underscore ___int64_t_defined in _default_types.h. newlib
renamed it, and a toolchain carrying a mismatched pair of those headers
still gets uint64_t but silently loses the whole PRI*64 block. Hence the
giveaway in the report: line 224 declares a uint64_t quite happily and
only line 226 fails.

Nothing to do with this repository, then, but there is no reason to
depend on that macro chain when we cannot test the toolchains people
build with. PRIu64 expands to "llu" on 32 bit ARM and uint64_t already
is unsigned long long, so this is the same format string and the same
argument - the built kernel.img and kernel.elf are byte for byte what
they were before the change. It was also the only PRI*/SCN* use in the
tree, uspi included.

Reported-and-fixed-by: Bill Howland

+++ src/rpi/performance.c
@@ -223,7 +223,14 @@ void print_performance_counters(perf_counters_t *pct) {
 	int i;
 	uint64_t cycle_counter = pct->cycle_counter;
 	cycle_counter *= 64;
-	printf("%26s = %"PRIu64"\r\n", "cycle counter", cycle_counter);
+	// Deliberately not PRIu64. Some arm-none-eabi newlib builds ship a
+	// mismatched sys/_stdint.h and machine/_default_types.h, which leaves
+	// __int64_t_defined unset - uint64_t still works, but inttypes.h quietly
+	// drops the whole PRI*64 block and this line fails with a baffling
+	// "expected ')' before 'PRIu64'". PRIu64 expands to "llu" on 32 bit ARM
+	// regardless, so this is the same format string with no toolchain
+	// dependency. Please do not "fix" it back.
+	printf("%26s = %llu\r\n", "cycle counter", (unsigned long long)cycle_counter);
 	for (i = 0; i < pct->num_counters; i++) {
 		printf("%26s = %u\r\n", type_lookup(pct->type[i]), pct->counter[i]);
 	}
